requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
bs4
arxiv
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
langchain_chroma
duckdb
langchain-groq
duckduckgo-search
openai
langchain-groq
streamlit
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
pypdf
faiss-cpu


